// lib/api/generated/models/createProjectTaskRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Heymoa API
 * Heymoa 서버 REST API
 * OpenAPI spec version: 1.0.0
 */
import type { CreateProjectTaskRequestAssignee } from "./createProjectTaskRequestAssignee";

export interface CreateProjectTaskRequest {
  /**
   * 기한 날짜. 없으면 null
   * @nullable
   */
  due?: string | null;
  /**
   * 할 일 담당. 사람 {type: USER|GUEST, id} 만 고릅니다. 없으면 null
   * @nullable
   */
  assignee?: CreateProjectTaskRequestAssignee;
  /**
   * 할 일 내용
   * @minLength 1
   */
  content: string;
}